90% improved memory efficiency
This was a slow process of profiling the code and removing unnecessary
functions until I found the culprit: I was pulling the “items” from
each members collection using xml_find_all and then using xml_find_all
on the result of that. Turns out, this not just redundant but
horrendously inefficient. This is probably an error introduced by my
early understanding of how the xml2 library worked.