Small optimization
Whenever you request the games list for a user, if it's changed since
the last query it gets queued. With a list 6000 users long, the stack
model can result be slow because users can update their list again well
before we get to them. Randomizing the stack every so often is a stupid
optimization, but it might improve performance just enough.